Rozdział2/requirements.txt
absl-py==0.2.2
astor==0.7.1
backcall==0.1.0
bleach==2.1.4
colorama==0.3.9
cycler==0.10.0
decorator==4.3.0
entrypoints==0.2.3
gast==0.2.0
graphviz==0.8.4
grpcio==1.13.0
h5py==2.8.0
hmmlearn==0.2.0
html5lib==1.0.1
ipykernel==4.8.2
ipython==6.5.0
ipython-genutils==0.2.0
jedi==0.12.1
Jinja2==2.10
joblib==0.12.2
jsonschema==2.6.0
jupyter-client==5.2.3
jupyter-core==4.4.0
Keras==2.2.0
Keras-Applications==1.0.2
Keras-Preprocessing==1.0.1
kiwisolver==1.0.1
livelossplot==0.2.0
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.2.2
mido==1.2.8
mistune==0.8.3
music21==5.1.0
nbconvert==5.3.1
nbformat==4.4.0
nltk==3.3
notebook==5.6.0
numpy==1.14.5
pandas==0.23.4
pandocfilters==1.4.2
parso==0.3.1
pickleshare==0.7.4
Pillow==5.2.0
pretty-midi==0.2.8
prometheus-client==0.3.1
prompt-toolkit==1.0.15
protobuf==3.6.0
pydot==1.2.4
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.7.3
pytz==2018.5
pywinpty==0.5.4
PyYAML==3.13
pyzmq==17.1.2
scikit-learn==0.19.2
scipy==1.1.0
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.11.0
sklearn==0.0
tensorboard==1.9.0
tensorflow==1.9.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.3.1
tornado==5.1
traitlets==4.3.2
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
